Extract logging setup into separate helper class
